Updated sidebar to remove ArticleCoreListCounter and handle the count in the ArticlesDrawer queryRenderer instead, renamed ProjectsListCounter to DrawerListCounter to make it easier to know what it refers to, prop sorting
